# docs/admin/openapi.json
{
  "components": {
    "schemas": {
...
      "AnnotationIn": {
        "properties": {
          "base_version": {
            "title": "Base Version",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "drop_clip": {
            "default": false,
            "title": "Drop Clip",
            "type": "boolean"
          },
          "needs_review": {
            "default": false,
            "title": "Needs Review",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "edited",
              "submitted"
            ],
            "title": "Status",
            "type": "string"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/Track"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "base_version",
          "tracks",
          "description",
          "status"
        ],
        "title": "AnnotationIn",
        "type": "object"
      },
...
      "AnnotationVersion": {
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "created_utc": {
            "format": "date-time",
            "title": "Created Utc",
            "type": "string"
          },
          "description_changed": {
            "title": "Description Changed",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tracks_count": {
            "title": "Tracks Count",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "status",
          "author",
          "created_utc",
          "tracks_count",
          "description_changed"
        ],
        "title": "AnnotationVersion",
        "type": "object"
      },
...
      "EventPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "total_capped": {
            "default": false,
            "title": "Total Capped",
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "EventPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublishRequest": {
        "properties": {
          "batch_name": {
            "pattern": "^[a-z0-9_]+$",
            "title": "Batch Name",
            "type": "string"
          }
        },
        "required": [
          "batch_name"
        ],
        "title": "PublishRequest",
        "type": "object"
      },
...
      "ReviewDecision": {
        "properties": {
          "decision": {
            "enum": [
              "accept",
              "reject"
            ],
            "title": "Decision",
            "type": "string"
          },
          "frame": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ReviewDecision",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Home Guard Admin Center",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/events": {
      "get": {
        "description": "Event history, newest first. `with_total=true` also fills `total` (counted up to 10000; beyond that\n`total` is 10000 and `total_capped` is true). `collection_id` keeps only events in that collection.",
        "operationId": "list_events_v1_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "site",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site"
            }
          },
          {
            "in": "query",
            "name": "customer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "in": "query",
            "name": "camera",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Camera"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "ai",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ai"
            }
          },
          {
            "in": "query",
            "name": "verdict",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verdict"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "from_utc",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Utc"
            }
          },
          {
            "in": "query",
            "name": "to_utc",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Utc"
            }
          },
          {
            "in": "query",
            "name": "reviewed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reviewed"
            }
          },
          {
            "in": "query",
            "name": "flagged",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Flagged"
            }
          },
          {
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter"
            }
          },
          {
            "in": "query",
            "name": "collection_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Collection Id"
            }
          },
          {
            "in": "query",
            "name": "would_raise",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Would Raise"
            }
          },
          {
            "in": "query",
            "name": "with_total",
            "required": false,
            "schema": {
              "default": false,
              "title": "With Total",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Events",
        "tags": [
          "events"
        ]
      }
    },
...
    "/v1/events/{event_id}/annotation": {
      "get": {
        "operationId": "get_annotation_v1_events__event_id__annotation_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Annotation",
        "tags": [
          "annotations"
        ]
      },
      "put": {
        "operationId": "save_annotation_v1_events__event_id__annotation_put",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Annotation",
        "tags": [
          "annotations"
        ]
      }
    },
    "/v1/events/{event_id}/annotation/history": {
      "get": {
        "operationId": "annotation_history_v1_events__event_id__annotation_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnnotationVersion"
                  },
                  "title": "Response Annotation History V1 Events  Event Id  Annotation History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Annotation History",
        "tags": [
          "annotations"
        ]
      }
    },
    "/v1/events/{event_id}/annotation/review": {
      "post": {
        "operationId": "review_annotation_v1_events__event_id__annotation_review_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Annotation",
        "tags": [
          "annotations"
        ]
      }
    },
...
    "/v1/studio/collections/{collection_id}/publish": {
      "post": {
        "operationId": "publish_collection_v1_studio_collections__collection_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "collection_id",
            "required": true,
            "schema": {
              "title": "Collection Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish Collection",
        "tags": [
          "annotations"
        ]
      }
    },
...
    "/v1/studio/publishes": {
      "get": {
        "operationId": "list_publishes_v1_studio_publishes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PublishOut"
                  },
                  "title": "Response List Publishes V1 Studio Publishes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Publishes",
        "tags": [
          "annotations"
        ]
      }
    },
...
  }
}